"update 3.1: self-defined serializer (choose from 4 serializers)"

+++ example-common/src/main/java/com/ypy/common/service/UserService.java
@@ -17,5 +17,5 @@ public interface UserService {
     /**
      * new method get number
      */
-    default short getNumber() { return 8964; }
+    default int getNumber() { return 8964; }
 }
